fastApiProject/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
python-multipart==0.0.9
pillow==10.2.0
opencv-python==4.9.0.80
torch==2.2.2
ultralytics==8.1.34
python-magic==0.4.27
requests==2.31.0
starlette==0.36.3
numpy==1.26.4
